+++ src/Chess/Pieces/BasePiece.java
@@ -111,7 +111,7 @@ public abstract class BasePiece extends MoveableImage{
         return distance <= movement_distance && !Engine.isOccupied_Tile(x,y);
     }
     public boolean can_capture_tile(int x,int y,int distance){
-        return distance <= Math.max(attack_distance,movement_distance) && Engine.isEnemy_Tile(x,y,this);
+        return distance <= attack_distance && Engine.isEnemy_Tile(x,y,this);
     }
 
     public List<Tile> getAvailable_movements(){
