+++ src/Chess/Pieces/Rook.java
@@ -9,6 +9,6 @@ public class Rook extends BasePiece{
     public Rook(Image image, Tile current_Tile,int ID) {
         super(image, current_Tile,"Rook",ID);
         this.movement_distance = 1;
-        this.attack_distance = 2;
+        this.attack_distance = 3;
     }
 }
